Fix azblob
